ECC-1930: Renamed: NumericLimits.h -> ecc_numeric_limits.h
