+++ src/ecc_numeric_limits.h
